winkernel: add more events to proxier_events
